# Improve marginal distribution