--- a/python/doc/user_manual/base_objects.rst
+++ b/python/doc/user_manual/base_objects.rst
@@ -116,6 +116,7 @@ Vectors
     CovarianceMatrixCollection
     ProcessCollection
     TestResultCollection
+    GaussKronrodRuleCollection
 
 Comparison operators
 ====================