… lib/test/CMakeLists.txt
… lib/test/t_MarginalDistribution_std.cxx
… lib/test/t_MarginalDistribution_std.expout
… lib/test/t_PosteriorDistribution_std.cxx
… lib/test/t_PosteriorDistribution_std.expout
… python/src/MarginalDistribution_doc.i.in
… python/test/CMakeLists.txt
… python/test/t_GaussKronrod_std.expout
… python/test/t_GaussKronrod_std.py
… python/test/t_IteratedQuadrature_std.py
… python/test/t_MarginalDistribution_std.expout
… python/test/t_MarginalDistribution_std.py
… python/test/t_PosteriorDistribution_std.expout
… python/test/t_PosteriorDistribution_std.py