This PR provides 2 significant improvements:
* The integration algorithms have been enhanced. It is now possible to better tune the 1D integration algorithm used by the IteratedQuadrature classe, and it is now possible to get the GaussKronrod rules by their name
* The MarginalDistribution.computePDF() method has been greatly improved. The special case where the marginal indices are a permutation of the underlying distribution is better taken into account, and if the usePDF_ flag is set to true (the default), then the PDF computation is done by integration over the marginalized components instead of using finite differences on the CDF. It results into a major speed-up, as seen [here](https://openturns.discourse.group/t/mutlivariate-distribution-with-combined-relationship/339/6). The test is reproduced here for convenience:

```python
import openturns as ot
import openturns.viewer as otv
from time import time

ot.Log.Show(ot.Log.ALL)
conditioningDist = ot.Normal(0.0, 1.0)
g = ot.SymbolicFunction(['y'], ['y', '0.1+y^2'])
conditionedDist = ot.Normal()
finalDist = ot.BayesDistribution(conditionedDist, conditioningDist, g)
modifyDist = finalDist.getMarginal([1, 0])

grid = ot.GridLayout(2, 3)

t0 = time()
g = finalDist.drawPDF([-1.0]*2, [1.0]*2)
g.setTitle("finalDist, t=%.3e" % (time() - t0) + "s")
grid.setGraph(0, 0, g)

t0 = time()
marginal = finalDist.getMarginal(0)
g = marginal.drawPDF(-1.0, 1.0)
g.setTitle("finalDist_0, t=%.3e" % (time() - t0) + "s")
grid.setGraph(0, 1, g)

t0 = time()
marginal = finalDist.getMarginal(1)
g = marginal.drawPDF(-1.0, 1.0)
g.setTitle("finalDist_1, t=%.3e" % (time() - t0) + "s")
grid.setGraph(0, 2, g)

t0 = time()
g = modifyDist.drawPDF([-1.0]*2, [1.0]*2)
g.setTitle("modifyDist, t=%.3e" % (time() - t0) + "s")
grid.setGraph(1, 0, g)

t0 = time()
marginal = modifyDist.getMarginal(0)
g = marginal.drawPDF(-1.0, 1.0)
g.setTitle("modifyDist_0, t=%.3e" % (time() - t0) + "s")
grid.setGraph(1, 1, g)

print(modifyDist)
t0 = time()
marginal = modifyDist.getMarginal(1)
g = marginal.drawPDF(-1.0, 1.0)
g.setTitle("modifyDist_1, t=%.3e" % (time() - t0) + "s")
grid.setGraph(1, 2, g)

view = otv.View(grid)
view.save("master.png")
```

I get with the current master:

![master](https://github.com/openturns/openturns/assets/8220933/78ce753f-50e9-465b-99cb-debd8076120f)

On the first row you see the 2D PDF of finalDist, followed by the PDF of its two marginal distributions. On the second row, you have the PDF of the permuted distribution as well as its marginals, which should be the same as the other distribution but swapped. You see 2 things:

* There is a lot of numerical noise on the 2D PDF, and a visible glitch on the first marginal of the permuted distribution
* The time taken to draw all these graphs is indicated in the title, and the 2D pdf took nearly 3h to be computed on my quite powerful laptop

With this PR I get:

![branch](https://github.com/openturns/openturns/assets/8220933/75ff083d-cfa3-487d-a005-01d5933daa2d)

and you see that the numerical noise has disapear and the computation time has been drastically reduced (yes, it is 0.04s instead of 8795s for the 2D PDF, and 0.2s for each marginal PDF instead of 0.042s and 3.47e-4s). The 1D marginal distributions are “true” marginal distributions while the 2D PDF is just a permutation of the initial distribution (and is now correctly detected and taken into account). It is why the 2D PDF is so fast to compute while there is an 1D integration for each evaluation of the PDF of the 1D marginal distributions.

It also work with discrete underlying distributions:
```python
import openturns as ot
import openturns.viewer as otv
from time import time

ot.Log.Show(ot.Log.ALL)
finalDist = ot.UserDefined(ot.Normal(2).getSample(25))
print(finalDist.getSupport())
modifyDist = ot.MarginalDistribution(finalDist, [1, 0])

grid = ot.GridLayout(2, 3)

t0 = time()
g = finalDist.drawPDF([-1.0]*2, [1.0]*2)
g.setTitle("finalDist, t=%.3e" % (time() - t0) + "s")
grid.setGraph(0, 0, g)

t0 = time()
marginal = finalDist.getMarginal(0)
g = marginal.drawPDF(-1.0, 1.0)
g.setTitle("finalDist_0, t=%.3e" % (time() - t0) + "s")
grid.setGraph(0, 1, g)

t0 = time()
marginal = finalDist.getMarginal(1)
g = marginal.drawPDF(-1.0, 1.0)
g.setTitle("finalDist_1, t=%.3e" % (time() - t0) + "s")
grid.setGraph(0, 2, g)

t0 = time()
g = modifyDist.drawPDF([-1.0]*2, [1.0]*2)
g.setTitle("modifyDist, t=%.3e" % (time() - t0) + "s")
grid.setGraph(1, 0, g)

t0 = time()
marginal = modifyDist.getMarginal(0)
g = marginal.drawPDF(-1.0, 1.0)
g.setTitle("modifyDist_0, t=%.3e" % (time() - t0) + "s")
grid.setGraph(1, 1, g)

print(modifyDist)
t0 = time()
marginal = modifyDist.getMarginal(1)
g = marginal.drawPDF(-1.0, 1.0)
g.setTitle("modifyDist_1, t=%.3e" % (time() - t0) + "s")
grid.setGraph(1, 2, g)

view = otv.View(grid)
view.save("branch_discrete.png")
```

giving:

![branch_discrete](https://github.com/openturns/openturns/assets/8220933/1a7b52cf-ba7f-430a-a3fa-14fa4657716a)

```