@jschueller 2024-04-04
great! it probably solves #2067, #2523