@mbaudin47 2024-04-16
@regislebrun : If you insert `python` in the header of your script written upper in this thread, just after the tripled backquotes, then it will be Python-colored. This is explained [here](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/creating-and-highlighting-code-blocks#syntax-highlighting) for ruby.