@jschueller 2024-04-26
this might close #2570 too